+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvThermalSourceTest.groovy
@@ -65,7 +65,6 @@ class CsvThermalSourceTest extends Specification implements CsvTestDataMeta {
       operationTime == sptd.thermalStorage.operationTime
       thermalBus == sptd.thermalStorage.thermalBus
       storageVolumeLvl == sptd.storageVolumeLvl
-      storageVolumeLvlMin == sptd.storageVolumeLvlMin
       inletTemp == sptd.inletTemp
       returnTemp == sptd.returnTemp
       c == sptd.c
@@ -84,7 +83,6 @@ class CsvThermalSourceTest extends Specification implements CsvTestDataMeta {
       operationTime == sptd.thermalStorage.operationTime
       thermalBus == sptd.thermalStorage.thermalBus
       storageVolumeLvl == sptd.storageVolumeLvl
-      storageVolumeLvlMin == sptd.storageVolumeLvlMin
       inletTemp == sptd.inletTemp
       returnTemp == sptd.returnTemp
       c == sptd.c